Configurable large font for dual time with analog clock
Dual time can be configured to use a bit larger font with analog clock.
If both dual time and altide are configured then smaller font is used
due to space restrictions. This setting really has an effect only in
round watches as in other systems the system font size is the same for
TINY and SMALL.

Fixed also a bug in time display.